Nix conditional on creation/mods to `vm_user_name`.
While testing this playbook 'for real' (i.e., on a pristine Ubuntu 20.04
installation and not using a podman container + molecule), I discovered
that it failed to change my shell from `/bin/bash` to `/usr/bin/zsh`.
This was due to my making the play for it conditional upon
`this_is_a_test` being true. I worried―for good reasonso―that this would
change the password of my current user. But the key insight was this:
it's not *necessary* for me to set a password during testing. And if I
do NOT set a password, I can set `shell: /usr/bin/zsh` and it will:

1. Create `vm_user_name` i.f.f. that user doesn't already exist, and:
2. Change the shell for `vm_user_name` to `/usr/bin/zsh` i.f.f. Zsh
   isn't already that user's shell.

Bottom line: the simpler version that doesn't specify a password and
isn't conditional on `this_is_a_test` being true has exactly the
required semantics.